# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TREstate — Turkish real estate aggregator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface LocationOut {
  id: number;
  level: string;
  name: string;
  slug: string;
}

export interface LocationNode {
  id: number;
  level: string;
  name: string;
  slug: string;
  children: LocationNode[];
}

export interface ImageOut {
  url: string;
  storage: string;
  position: number;
}

export type PriceInfoConverted = {[key: string]: number};

export interface PriceInfo {
  /** @nullable */
  initial_price?: number | null;
  /** @nullable */
  current_price?: number | null;
  currency?: string;
  /** @nullable */
  change_abs?: number | null;
  /** @nullable */
  change_pct?: number | null;
  price_changes_count?: number;
  /** @nullable */
  last_change_at?: string | null;
  days_in_db?: number;
  converted?: PriceInfoConverted;
}

export interface FreshnessInfo {
  first_seen_at: string;
  last_seen_at: string;
  last_checked_at: string;
  /** @nullable */
  last_changed_at?: string | null;
  status: string;
}

export interface ListingCard {
  id: number;
  transaction_type: string;
  property_type: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  price?: number | null;
  currency: string;
  /** @nullable */
  price_per_gross_m2?: number | null;
  /** @nullable */
  price_per_net_m2?: number | null;
  /** @nullable */
  gross_area_m2?: number | null;
  /** @nullable */
  net_area_m2?: number | null;
  /** @nullable */
  rooms?: string | null;
  /** @nullable */
  floor?: number | null;
  /** @nullable */
  total_floors?: number | null;
  /** @nullable */
  building_age?: number | null;
  /** @nullable */
  furnished?: boolean | null;
  /** @nullable */
  pool?: boolean | null;
  /** @nullable */
  balcony?: boolean | null;
  /** @nullable */
  parking?: boolean | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  status: string;
  first_seen_at: string;
  last_checked_at: string;
  province?: LocationOut | null;
  district?: LocationOut | null;
  neighbourhood?: LocationOut | null;
  images: ImageOut[];
}

export type ListingDetailFeatures = {[key: string]: string | null};

export type ListingDetail = ListingCard & ({
  /** @nullable */
  description?: string | null;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  heating?: string | null;
  /** @nullable */
  elevator?: boolean | null;
  /** @nullable */
  residential_complex?: string | null;
  /** @nullable */
  deed_status?: string | null;
  /** @nullable */
  distance_to_sea_m?: number | null;
  /** @nullable */
  address_text?: string | null;
  /** @nullable */
  seller_type?: string | null;
  /** @nullable */
  agency_name?: string | null;
  /** @nullable */
  publication_date?: string | null;
  source_code: string;
  external_id: string;
  /** @nullable */
  original_url?: string | null;
  price_info?: PriceInfo | null;
  freshness?: FreshnessInfo | null;
  features?: ListingDetailFeatures;
});

export interface ListingPage {
  items: ListingCard[];
  total: number;
  page: number;
  per_page: number;
}

/**
 * @nullable
 */
export type HistoryPointChangedFields = { [key: string]: unknown } | null;

export interface HistoryPoint {
  observed_at: string;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  gross_area_m2?: number | null;
  /** @nullable */
  net_area_m2?: number | null;
  /** @nullable */
  changed_fields?: HistoryPointChangedFields;
}

export interface MapPoint {
  id: number;
  latitude: number;
  longitude: number;
  /** @nullable */
  price?: number | null;
  currency: string;
  /** @nullable */
  rooms?: string | null;
}

export type MetaResponseExchangeRates = {[key: string]: {
  rate_to_try?: number;
  fetched_at?: string;
}};

export interface MetaResponse {
  transaction_types: string[];
  property_types: string[];
  currencies: string[];
  exchange_rates: MetaResponseExchangeRates;
  sorts: string[];
}

export interface FavoriteInput {
  listing_id: number;
}

export type GetListingsParams = {
/**
 * @nullable
 */
transaction_type?: string | null;
/**
 * @nullable
 */
property_type?: string | null;
/**
 * @nullable
 */
district?: string | null;
/**
 * @nullable
 */
neighbourhood?: string | null;
rooms?: string[];
/**
 * @nullable
 */
min_price?: number | null;
/**
 * @nullable
 */
max_price?: number | null;
/**
 * @nullable
 */
min_m2?: number | null;
/**
 * @nullable
 */
max_m2?: number | null;
/**
 * @nullable
 */
furnished?: boolean | null;
/**
 * @nullable
 */
pool?: boolean | null;
/**
 * @nullable
 */
parking?: boolean | null;
/**
 * @nullable
 */
balcony?: boolean | null;
only_price_drops?: boolean;
sort?: GetListingsSort;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
per_page?: number;
};

export type GetListingsSort = typeof GetListingsSort[keyof typeof GetListingsSort];


export const GetListingsSort = {
  newest: 'newest',
  oldest: 'oldest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  ppm2_asc: 'ppm2_asc',
} as const;

export type GetMapPointsParams = {
/**
 * @nullable
 */
transaction_type?: string | null;
/**
 * @nullable
 */
property_type?: string | null;
/**
 * @nullable
 */
district?: string | null;
/**
 * @nullable
 */
min_lat?: number | null;
/**
 * @nullable
 */
max_lat?: number | null;
/**
 * @nullable
 */
min_lon?: number | null;
/**
 * @nullable
 */
max_lon?: number | null;
/**
 * @maximum 2000
 */
limit?: number;
};