# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TREstate — Turkish real estate aggregator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Search listings
      description: Search and filter property listings with pagination
      parameters:
        - name: transaction_type
          in: query
          schema:
            type: ["string", "null"]
        - name: property_type
          in: query
          schema:
            type: ["string", "null"]
        - name: district
          in: query
          schema:
            type: ["string", "null"]
        - name: neighbourhood
          in: query
          schema:
            type: ["string", "null"]
        - name: rooms
          in: query
          schema:
            type: array
            items:
              type: string
          explode: true
          style: form
        - name: min_price
          in: query
          schema:
            type: ["number", "null"]
        - name: max_price
          in: query
          schema:
            type: ["number", "null"]
        - name: min_m2
          in: query
          schema:
            type: ["number", "null"]
        - name: max_m2
          in: query
          schema:
            type: ["number", "null"]
        - name: furnished
          in: query
          schema:
            type: ["boolean", "null"]
        - name: pool
          in: query
          schema:
            type: ["boolean", "null"]
        - name: parking
          in: query
          schema:
            type: ["boolean", "null"]
        - name: balcony
          in: query
          schema:
            type: ["boolean", "null"]
        - name: only_price_drops
          in: query
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          schema:
            type: string
            default: newest
            enum: [newest, oldest, price_asc, price_desc, ppm2_asc]
        - name: page
          in: query
          schema:
            type: number
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: number
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingPage"
  /listings/{listing_id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing detail
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
  /listings/{listing_id}/history:
    get:
      operationId: getListingHistory
      tags: [listings]
      summary: Get listing price history
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Price history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryPoint"
...
components:
  schemas:
...
    LocationOut:
      type: object
      properties:
        id:
          type: number
        level:
          type: string
        name:
          type: string
        slug:
          type: string
      required: [id, level, name, slug]
...
    ImageOut:
      type: object
      properties:
        url:
          type: string
        storage:
          type: string
        position:
          type: number
      required: [url, storage, position]
    PriceInfo:
      type: object
      properties:
        initial_price:
          type: ["number", "null"]
        current_price:
          type: ["number", "null"]
        currency:
          type: string
        change_abs:
          type: ["number", "null"]
        change_pct:
          type: ["number", "null"]
        price_changes_count:
          type: number
        last_change_at:
          type: ["string", "null"]
          format: date-time
        days_in_db:
          type: number
        converted:
          type: object
          additionalProperties:
            type: number
    FreshnessInfo:
      type: object
      properties:
        first_seen_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        last_checked_at:
          type: string
          format: date-time
        last_changed_at:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
      required: [first_seen_at, last_seen_at, last_checked_at, status]
    ListingCard:
      type: object
      properties:
        id:
          type: number
        transaction_type:
          type: string
        property_type:
          type: string
        title:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: string
        price_per_gross_m2:
          type: ["number", "null"]
        price_per_net_m2:
          type: ["number", "null"]
        gross_area_m2:
          type: ["number", "null"]
        net_area_m2:
          type: ["number", "null"]
        rooms:
          type: ["string", "null"]
        floor:
          type: ["number", "null"]
        total_floors:
          type: ["number", "null"]
        building_age:
          type: ["number", "null"]
        furnished:
          type: ["boolean", "null"]
        pool:
          type: ["boolean", "null"]
        balcony:
          type: ["boolean", "null"]
        parking:
          type: ["boolean", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
        first_seen_at:
          type: string
          format: date-time
        last_checked_at:
          type: string
          format: date-time
        province:
          oneOf:
            - $ref: "#/components/schemas/LocationOut"
            - type: "null"
        district:
          oneOf:
            - $ref: "#/components/schemas/LocationOut"
            - type: "null"
        neighbourhood:
          oneOf:
            - $ref: "#/components/schemas/LocationOut"
            - type: "null"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ImageOut"
      required:
        - id
        - transaction_type
        - property_type
        - currency
        - status
        - first_seen_at
        - last_checked_at
        - images
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/ListingCard"
        - type: object
          properties:
            description:
              type: ["string", "null"]
            bedrooms:
              type: ["number", "null"]
            bathrooms:
              type: ["number", "null"]
            heating:
              type: ["string", "null"]
            elevator:
              type: ["boolean", "null"]
            residential_complex:
              type: ["string", "null"]
            deed_status:
              type: ["string", "null"]
            distance_to_sea_m:
              type: ["number", "null"]
            address_text:
              type: ["string", "null"]
            seller_type:
              type: ["string", "null"]
            agency_name:
              type: ["string", "null"]
            publication_date:
              type: ["string", "null"]
              format: date
            source_code:
              type: string
            external_id:
              type: string
            original_url:
              type: ["string", "null"]
            price_info:
              oneOf:
                - $ref: "#/components/schemas/PriceInfo"
                - type: "null"
            freshness:
              oneOf:
                - $ref: "#/components/schemas/FreshnessInfo"
                - type: "null"
            features:
              type: object
              additionalProperties:
                type: ["string", "null"]
          required: [source_code, external_id]
    ListingPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListingCard"
        total:
          type: number
        page:
          type: number
        per_page:
          type: number
      required: [items, total, page, per_page]
    HistoryPoint:
      type: object
      properties:
        observed_at:
          type: string
          format: date-time
        price:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        gross_area_m2:
          type: ["number", "null"]
        net_area_m2:
          type: ["number", "null"]
        changed_fields:
          type: ["object", "null"]
          additionalProperties: true
      required: [observed_at]
...