# lib/api-zod/src/generated/types/freshnessInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TREstate — Turkish real estate aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface FreshnessInfo {
  first_seen_at: Date;
  last_seen_at: Date;
  last_checked_at: Date;
  /** @nullable */
  last_changed_at?: Date | null;
  status: string;
}