# lib/api-zod/src/generated/types/historyPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TREstate — Turkish real estate aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryPointChangedFields } from './historyPointChangedFields';

export interface HistoryPoint {
  observed_at: Date;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  gross_area_m2?: number | null;
  /** @nullable */
  net_area_m2?: number | null;
  /** @nullable */
  changed_fields?: HistoryPointChangedFields;
}