# lib/api-zod/src/generated/types/imageOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TREstate — Turkish real estate aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface ImageOut {
  url: string;
  storage: string;
  position: number;
}