# lib/api-zod/src/generated/types/listingDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TREstate — Turkish real estate aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { FreshnessInfo } from './freshnessInfo';
import type { ListingCard } from './listingCard';
import type { ListingDetailFeatures } from './listingDetailFeatures';
import type { PriceInfo } from './priceInfo';

export type ListingDetail = ListingCard & ({
  /** @nullable */
  description?: string | null;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  heating?: string | null;
  /** @nullable */
  elevator?: boolean | null;
  /** @nullable */
  residential_complex?: string | null;
  /** @nullable */
  deed_status?: string | null;
  /** @nullable */
  distance_to_sea_m?: number | null;
  /** @nullable */
  address_text?: string | null;
  /** @nullable */
  seller_type?: string | null;
  /** @nullable */
  agency_name?: string | null;
  /** @nullable */
  publication_date?: Date | null;
  source_code: string;
  external_id: string;
  /** @nullable */
  original_url?: string | null;
  price_info?: PriceInfo | null;
  freshness?: FreshnessInfo | null;
  features?: ListingDetailFeatures;
});