# lib/api-zod/src/generated/types/listingPage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TREstate — Turkish real estate aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingCard } from './listingCard';

export interface ListingPage {
  items: ListingCard[];
  total: number;
  page: number;
  per_page: number;
}