# lib/api-zod/src/generated/types/locationOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TREstate — Turkish real estate aggregator API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationOut {
  id: number;
  level: string;
  name: string;
  slug: string;
}